+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Windows is nonportable if the EXE path contains `\Program Files`
   - Linux is nonportable if EXE path contains `/bin/`, `/usr/`, or `/.local/bin/`
 - Moved some view settings from "Geometry" to a new section "Culling"
+- Animation logic for very fast twist speeds
 
 ### Fixed